+++ model/src/main/java/com/idiot/model/EstateListBrokerResponse.kt
@@ -1,7 +1,7 @@
 package com.idiot.model
 
 data class EstateListBrokerResponse(
-  val estateId: Int,
+  val estateId: Long,
   val ownerName: String,
   val ownerPhone: String,
   val estateAddress: RegisterEstateOwner
